# src/api/company/dto/UpdateCompanyDto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString, MaxLength } from 'class-validator';

export class UpdateCompanyDto {
  @ApiProperty({ example: 'Acme Corporation' })
  @IsString()
  @IsNotEmpty()
  @MaxLength(200)
  name!: string;

  @ApiProperty({ example: 'cmp_live_new_key' })
  @IsString()
  @IsNotEmpty()
  @MaxLength(128)
  apiKey!: string;
}